# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MenuAI – AI-powered menu translation and allergen platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menus/{menuId}/sections:
    post:
      operationId: createSection
      tags: [sections]
      summary: Create a section in a menu
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionBody"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /sections/{id}:
    patch:
      operationId: updateSection
      tags: [sections]
      summary: Update a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectionBody"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dishes ───────────────────────────────────────────────────
  /sections/{sectionId}/dishes:
    post:
      operationId: createDish
      tags: [dishes]
      summary: Create a dish in a section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDishBody"
      responses:
        "201":
          description: Created dish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
  /dishes/{id}:
    get:
      operationId: getDish
      tags: [dishes]
      summary: Get a dish with its allergens, labels, and translations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dish detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DishDetail"
        "404":
          description: Not found
    patch:
      operationId: updateDish
      tags: [dishes]
      summary: Update a dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDishBody"
      responses:
        "200":
          description: Updated dish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
    delete:
      operationId: deleteDish
      tags: [dishes]
      summary: Delete a dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dishes/{id}/availability:
    patch:
      operationId: toggleDishAvailability
      tags: [dishes]
      summary: Toggle dish availability (86 a dish or restore it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleDishAvailabilityBody"
      responses:
        "200":
          description: Updated dish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"

  # ── Allergens ────────────────────────────────────────────────
  /dishes/{dishId}/allergens:
    post:
      operationId: addAllergen
      tags: [allergens]
      summary: Add an allergen to a dish
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAllergenBody"
      responses:
        "201":
          description: Created allergen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allergen"
  /allergens/{id}:
    patch:
      operationId: updateAllergen
      tags: [allergens]
      summary: Confirm or override an allergen flag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllergenBody"
      responses:
        "200":
          description: Updated allergen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allergen"
    delete:
      operationId: deleteAllergen
      tags: [allergens]
      summary: Remove an allergen from a dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Analytics (wow endpoints) ─────────────────────────────────
...
components:
  schemas:
...
    Section:
      type: object
      properties:
        id:
          type: integer
        menuId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, menuId, name, sortOrder, createdAt]
...
    CreateSectionBody:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
      required: [name]
    UpdateSectionBody:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer

    # ── Dish ───────────────────────────────────────────────────
    Dish:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
          default: "KES"
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
          default: true
        isSpecial:
          type: boolean
          default: false
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sectionId, name, currency, isAvailable, isSpecial, sortOrder, createdAt, updatedAt]
    DishDetail:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isSpecial:
          type: boolean
        sortOrder:
          type: integer
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/Allergen"
        dietaryLabels:
          type: array
          items:
            type: string
        translations:
          type: array
          items:
            $ref: "#/components/schemas/DishTranslation"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sectionId, name, currency, isAvailable, isSpecial, sortOrder, allergens, dietaryLabels, translations, createdAt, updatedAt]
    CreateDishBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
          default: "KES"
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
          default: true
        isSpecial:
          type: boolean
          default: false
        sortOrder:
          type: integer
        dietaryLabels:
          type: array
          items:
            type: string
      required: [name]
    UpdateDishBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isSpecial:
          type: boolean
        sortOrder:
          type: integer
        dietaryLabels:
          type: array
          items:
            type: string
    ToggleDishAvailabilityBody:
      type: object
      properties:
        isAvailable:
          type: boolean
      required: [isAvailable]

    # ── Allergen ───────────────────────────────────────────────
    Allergen:
      type: object
      properties:
        id:
          type: integer
        dishId:
          type: integer
        allergenType:
          type: string
          enum: [gluten, dairy, nuts, peanuts, eggs, soy, shellfish, fish, sesame, sulphites, celery, mustard, lupin, molluscs]
        isAiSuggested:
          type: boolean
        isConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, dishId, allergenType, isAiSuggested, isConfirmed, createdAt]
    AddAllergenBody:
      type: object
      properties:
        allergenType:
          type: string
          enum: [gluten, dairy, nuts, peanuts, eggs, soy, shellfish, fish, sesame, sulphites, celery, mustard, lupin, molluscs]
        isAiSuggested:
          type: boolean
          default: false
        isConfirmed:
          type: boolean
          default: true
      required: [allergenType]
    UpdateAllergenBody:
      type: object
      properties:
        isConfirmed:
          type: boolean

    # ── Translation ────────────────────────────────────────────
    DishTranslation:
      type: object
      properties:
        id:
          type: integer
        dishId:
          type: integer
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        culinaryExplanation:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, dishId, languageCode, name, createdAt]

    # ── Analytics ──────────────────────────────────────────────
...